requirements.txt
crewai 
langchain 
chromadb 
fastapi 
uvicorn 
groq 
langchain-groq 
python-dotenv
chromadb 
sentence-transformers